6-litellm-proxy/pyproject.toml
[project]
name = "6-litellm-proxy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.36.0",
    "langfuse>=3.3.4",
    "litellm[proxy]>=1.76.1",
    "lunary>=1.4.27",
    "python-dotenv>=1.1.1",
]
